+++ packages/cli/src/commands/gateway/configure.ts
@@ -1,5 +1,3 @@
-import { configureGateway, reconcileServices } from '@denvig/sdk/unsafe'
-
 import { Command } from '../../lib/command.ts'
 
 export const gatewayConfigureCommand = new Command({
@@ -10,9 +8,9 @@ export const gatewayConfigureCommand = new Command({
   example: 'gateway configure',
   args: [],
   flags: [],
-  handler: async ({ flags }) => {
-    const reconcileResult = await reconcileServices()
-    const result = await configureGateway()
+  handler: async ({ sdk, flags }) => {
+    const { reconcile: reconcileResult, gateway: result } =
+      await sdk.gateway.configure()
 
     if (!result) {
       console.error(